started moving fetches to thier own files
